+++ insta/src/macros.rs
@@ -278,7 +278,7 @@ macro_rules! _prepare_snapshot_for_redaction {
 #[macro_export]
 macro_rules! _prepare_snapshot_for_redaction {
     ($value:expr, {$($k:expr => $v:expr),*}, $format:ident) => {
-        compile_error!("insta was compiled without redaction support.");
+        "insta was compiled without redactions support. Enable the `redactions` feature."
     };
 }
 
@@ -293,6 +293,20 @@ macro_rules! _prepare_snapshot_for_redaction {
 macro_rules! assert_debug_snapshot {
     ($($arg:tt)*) => {
         $crate::_assert_snapshot_base!(transform=|v| std::format!("{:#?}", v), $($arg)*)
+    };
+}
+
+/// Asserts a `Debug` snapshot in compact format.
+///
+/// The value needs to implement the `fmt::Debug` trait.  This is useful for
+/// simple values that do not implement the `Serialize` trait, but does not
+/// permit redactions.
+///
+/// Debug is called with `"{:?}"`, which means this does not use pretty-print.
+#[macro_export]
+macro_rules! assert_compact_debug_snapshot {
+    ($($arg:tt)*) => {
+        $crate::_assert_snapshot_base!(transform=|v| std::format!("{:?}", v), $($arg)*)
     };
 }