+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to insta and cargo-insta are documented here.
 
 ## 1.40.0
 
+- Inline snapshots now use the required number of `#`s to escape the snapshot
+  value, rather than always using `###`. This allows snapshotting values which
+  themselves contain `###`. If there are no existing `#` characters in the
+  snapshot value, a single `#` will be used.
+
+- Print a warning when encountering old snapshot formats.  #503
+
 - No longer suggest running `cargo insta` message when running `cargo insta test --check`.  #515
 
 - Print a clearer error message when accepting a snapshot that was removed.  #516